tests: Move `GET /crates` tests to `krate::search` module
